Separate file management script files.
