+++ src/data_exploration/explore.py
@@ -9,6 +9,8 @@ import colorsys
 import zipfile
 import os
 import cv2
+from src.manage_files import *
+from src.livecell import LiveCellSet
 
 ### Dataset sources ###
 # MEDIAR
@@ -25,124 +27,10 @@ import cv2
 # CellPose
 # https://www.cellpose.org/dataset
 # 
-# LiveCell
-# -- NOT -- https://www.kaggle.com/competitions/sartorius-cell-instance-segmentation/data
-# https://github.com/sartorius-research/LIVECell
-# http://livecell-dataset.s3.eu-central-1.amazonaws.com/LIVECell_dataset_2021/images.zip
-# 
 # DataScienceBowl
 # https://www.kaggle.com/competitions/data-science-bowl-2018/data
 #
 ###
-
-VISIBLE_TYPES = [".bmp", ".png"] # + [".jpg", ".jpeg"]
-TENSOR_TYPES = [".tif", ".tiff"]
-IMAGE_TYPES = VISIBLE_TYPES + TENSOR_TYPES
-MISC_TYPES = [".md", ".zip", ".txt", ".csv", ".py", ".json", ""]
-FILE_TYPES = IMAGE_TYPES + MISC_TYPES
-
-LABELED = 'Labeled'
-MASK = 'Mask'
-UNLABELED = 'Unlabeled'
-SYNTHETIC = 'Synthetic'
-
-def split_filepath(filepath):
-    dirpath, filename = os.path.split(filepath)
-    name, ext = os.path.splitext(filename)
-    separator = '/' if (not dirpath or dirpath[-1] != '/') else ''
-    return dirpath + separator, name, ext
-
-def unzip_archive(root, filepath):
-    dirpath, name, ext = split_filepath(filepath)
-    print("Inspecting archive: ", filepath)
-    with zipfile.ZipFile(root + filepath, 'r') as zip_ref:
-        # Top level folders
-        folders = set(dirpath + zipname.split('/')[0] for zipname in zip_ref.namelist())
-        missing = [folder for folder in folders if not os.path.exists(root + folder)]
-        if missing:
-            print("Unzipping archive: ", missing)
-            zip_ref.extractall(root + dirpath)
-            return missing
-        return []
-
-def list_dataset(root, folder = '/'):
-    files_by_type = {type:set() for type in FILE_TYPES}
-    for filepath in enumerate_dataset(root, folder):
-        dirpath, name, ext = split_filepath(filepath)
-        files_by_type[ext].add(filepath)
-    return files_by_type
-
-def unzip_dataset(root, folder):
-    for filepath in enumerate_dataset(root, folder):
-        unzip_datafile(root, filepath)
-
-def unzip_datafile(root, filepath):
-    dirpath, name, ext = split_filepath(filepath)
-    if ext != ".zip":
-        return
-    for unzipped in unzip_archive(root, filepath):
-        if not os.path.exists(root + unzipped):
-            print("!WARNING! Archive did not produce folder: ", root + unzipped)
-        elif os.path.isdir(root + unzipped):
-                unzip_dataset(root, unzipped)
-        else:
-            unzip_datafile(root, unzipped)
-
-def enumerate_dataset(root, folder):
-    #print("Enumerating folder: ", folder)
-    for filename in os.listdir(root + folder):
-        filepath = folder + filename
-        yield filepath
-        dirpath, name, ext = split_filepath(filepath)
-        if os.path.isdir(root + filepath):
-            for fullpath in enumerate_dataset(root, filepath + "/"):
-                yield fullpath
-
-class BaseFileSet:
-    def __init__(self, root):
-        self.root = root
-        self.logs = []
-
-    def blacklist(self, filepath):
-        return False
-    
-    def signature(self, filepath):
-        (dims, channels, type) = tensor_signature(self.load(filepath))
-        category = self.categorize(filepath)
-        return (dims, channels, type, category)
-    
-    def enumerate(self, folder):
-        for filepath in enumerate_dataset(folder, self.root + "/"):
-            dirpath, name, ext = split_filepath(filepath)
-            if ext not in IMAGE_TYPES:
-                continue
-            if self.blacklist(filepath):
-                continue
-            yield filepath
-    
-    def mask_filepath(self, filepath):
-        return []
-    
-    def categorize(self, filepath):
-        return UNLABELED
-    
-    def load(self, filepath):
-        return load_image(filepath)
-    
-class LiveCellSet(BaseFileSet):
-    def blacklist(self, filepath):
-        return "/LIVECell_dataset_2021" not in filepath
-    
-    def mask_filepath(self, filepath):
-        if "/images" in filepath:
-            yield (filepath.replace("/images", "/masks"), MASK)
-
-    def categorize(self, filepath):
-        if "/masks" in filepath:
-            return MASK
-        if "/images" in filepath:
-            return LABELED
-        return UNLABELED
 
 class ScienceBowlSet(BaseFileSet):
     def mask_filepath(self, filepath):
@@ -386,22 +274,6 @@ def save_maskframe(mask_path, frame_path):
     maskframe = mask_frame(tensor)
     maskframe.to_csv(frame_path)
 
-def target_file(filepath, ext):
-    dirpath, name, _ = split_filepath(filepath)
-    return dirpath + name + ext
-
-def safely_process(log, process, overwrite=False):
-    def wrapper(source, target):
-        try:
-            if overwrite or not os.path.exists(target):
-                dirpath, _, _ = split_filepath(target)
-                os.makedirs(dirpath, exist_ok=True)
-                process(source, target)
-        except Exception as e:
-            print(e)
-            log.append(source + " -> " + target)
-    return wrapper
-
 # Save black-white mask
 def save_bw_mask(source, target):
     imgT = DataSet.load_raw(None, source)
@@ -449,20 +321,6 @@ def save_clean_image(source, target):
     img = Image.fromarray(img.astype('uint8'))
     img.save(target)
     #cv2.imwrite(target, img)
-
-def load_image(img_path):
-    dirpath, name, ext = split_filepath(img_path)
-    if ext in TENSOR_TYPES:
-        return tif.imread(img_path)
-    elif ext in VISIBLE_TYPES:
-        return io.imread(img_path)
-
-def save_image(img_path, image):
-    dirpath, name, ext = split_filepath(img_path)
-    if ext in TENSOR_TYPES:
-        return tif.imwrite(img_path, image)
-    elif ext in VISIBLE_TYPES:
-        return io.imsave(img_path, image, check_contrast=False)
 
 def prepare_metaframe(dataroot, target_path):
     dataset = [ZenodoNeurIPS('/neurips'), CellposeSet("/cellpose"), OmniPoseSet("/omnipose"), LiveCellSet("/livecell"), ScienceBowlSet("/sciencebowl")]
@@ -673,11 +531,6 @@ def sanity_check_box(width, height, bdf):
     assert 0 <= bdf['Top'] <= height
     assert 0 <= bdf['Bottom'] <= height
 
-def tensor_signature(tensor):
-    dims = len(tensor.shape)
-    channels = 1 if dims < 3 else tensor.shape[-1]
-    return (dims, channels, tensor.dtype)
-
 def check_signatures(dataroot, datasets):
     for dataset in datasets:
         files = list(dataset.enumerate(dataroot))
@@ -711,6 +564,8 @@ def upper_bound(tensor):
 def sanitize(tensor, rescale):
     assert 0 <= tensor.min(), "Tensor contains negative values"
     bits = upper_bound(tensor)
+    if tensor.shape[-1] == 4:  # RGBA
+        tensor = color.rgba2rgb(tensor)
     if rescale and (bits != 8):
         scaled = tensor * ((1 << 8)/(1 << bits))
         return scaled.astype('uint8')
@@ -728,7 +583,7 @@ def unify_load(dataset, filepath):
 
 if __name__ == "__main__":
     dataroot = "./data"
-    datasets = [ZenodoNeurIPS("/neurips"), CellposeSet("/cellpose"), OmniPoseSet("/omnipose"), LiveCellSet("/livecell"), ScienceBowlSet("/sciencebowl")]
+    datasets = [ZenodoNeurIPS("/neurips"), CellposeSet("/cellpose"), OmniPoseSet("/omnipose"), LiveCellSet(), ScienceBowlSet("/sciencebowl")]
     #unzip_dataset(dataroot, "/raw/")
     #check_signatures(dataroot + "/raw", datasets[0:1])
 
@@ -745,6 +600,7 @@ if __name__ == "__main__":
     unify_dataset(dataroot, datasets[0])
     unify_dataset(dataroot, datasets[1])
     unify_dataset(dataroot, datasets[2])
+    unify_dataset(dataroot, datasets[3])
 
 if False: #__name__ == "__main__":
     dataset = DataSet("./data")
